requirements.txt
click>=8.0.0
pathlib2>=2.3.0; python_version < '3.4'

# Feature extraction
Pillow>=8.0.0  # Image metadata extraction

# Optional ML dependencies (for future clustering engine)
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0

# Optional LLM dependencies (for future LLM engine)
# openai>=1.0.0  # Uncomment when implementing LLM features
# langchain>=0.1.0  # Uncomment when implementing LLM features

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# GUI dependencies (optional)
streamlit>=1.20.0  # For web GUI

# Documentation
mkdocs>=1.4.0
mkdocs-material>=8.0.0